ExampleFiles/requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
Django==3.0.5
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-taggit==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
Pillow==7.1.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
Werkzeug==1.0.1
wrapt==1.11.2
